Add files via upload
Ultimas validaciones

+++ TP_1_CascaraNew/TP_1_Cascara/funciones.h
@@ -9,6 +9,6 @@ float multiply(float num1, float num2);
 
 float division(float num1, float num2);
 
-long long factoring(long long num1);
+int factoring(int num1);
 
 #endif // FUNCIONES_H_INCLUDED
